Update make_sym_links and README
Add vundle clone command to make_sym_links script. This will allow
the user to run :BundleInstall on the first use of vim and have all
the current versions of the plugins downloaded and installed. Also
add echo to script to provide additional instructions.

Edit README to correct typo in installation instructions as well as
misc revisions to body text.