Add UltiSnips to vimrc. Add .gitignore
Add a .gitignore file. Currently ignores the contents of the bundle
sub folder in the vim folder.

Add UltiSnips to vimrc, also added a mySnippets folder in the vim folder
this holds my personal snippets this is linked in the vimrc as well.
Starting off with a few html snippets and php snippets.