Modify config_linux_enviornment.sh
Modify config_linux_enviornment.sh to install nodejs from github
so we have a current version which will properly run coffee script.

Modify config_linux_enviornment to write progress to a log file.
This way if there is a problem with the install we can see where it
failed. The script prints the log file at the end to give the user
a summary of what has been installed. The log file is then deleted.