Enhancement: Enhanced yank to register
The enhanced yank was an overall win. However, it hosed yanking to a
named register in visual select mode. The best I have come up with is to
add another function that will provide this functionality. Now you can
do a <Leader>" and it will ask you which register to yank the visual
selection to.