Enhancement: Unite now uses git for the source of find file options.
The async file search only seems to cause pain and is virtually
impossible to debug. The source for finding files has now been switched
to use git. It is faster and more reliable than using async. The only
small concession is that it will not return files that have not yet been
staged.