Feature: Now using `Plug` as Vim plug-in manager
The previous plug-in manager was consistently unpredictable and recently
completely nonfunctional on my computer.  As a result I have replaced it
with the much more popular and straightforward `Plug` tool.

What this means for you. If you have not added any additional
plug-ins then this change should be transparent. If you have added
plug-ins in your `custom-configs` directory then you will need to make a
minor but crucial adjustment to those files.

Your plugin configuration files now must wrap the plug-in source in
`'` as shown below.

```
"===============================================================================
" Plugin source
"===============================================================================
"'ownerName/example'

"===============================================================================
" Plugin Configurations
"===============================================================================
" N/A
```

This is a necessary inconvenience that will hopefully increase stability
going forward. If you run into any problems open an issue and we will
get it sorted out.

Jarrod